bug(feedback):User-Interface Background color of navigation closing of the navbar Updating name in navbar Hamburger staying on place for very most of page(to continue working on) Horizontal scrolling is being removed making the footer a component solving a clickable issue on loging page putting components in to files and giving them their css
