+++ portfolio/src/components/Blog/Blog.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import Action from "./Actions";
-import blogPicture from "../pictures/6a01157041f4e3970b01b7c82eb758970b-320wi.png";
+import "./Blog.css";
+import Buttons from "../Buttons/Buttons";
+import blogPicture from "../../pictures/6a01157041f4e3970b01b7c82eb758970b-320wi.png";
 
 const Blog = (props) => {
     return (
@@ -9,7 +10,7 @@ const Blog = (props) => {
              <img  className="image" src={blogPicture} id="imageBlog" alt="blogPicture"></img>
             <p className="title">{props.title}</p>  
             <p className="script">{props.script}</p> 
-            <Action /> 
+            <Buttons /> 
             </div>
         </div>
     )
