+++ portfolio/src/components/Buttons/Buttons.css
@@ -1,0 +1,36 @@
+*{
+  margin: 0px;
+  border: 0px;
+}
+body{
+  background-color: black;
+  overflow: hidden;
+  overflow-y: auto;
+}
+
+.action{
+  display: flex;
+  justify-content: center;
+  justify-content: space-around;
+  padding-bottom: 1em;
+}
+
+.edit{
+  background-color: green;
+  color: black;
+  font-weight: bolder;    
+  border-radius: .5em;  
+  margin-top: 20px;
+  padding: 10px 25px 10px 25px;
+  cursor: pointer;
+}
+
+.delete{
+  background-color: red;
+  color: black;
+  font-weight: bolder;    
+  border-radius: .5em;  
+  margin-top: 20px;
+  padding: 7px 20px 7px 20px;
+  cursor: pointer;
+}
