+++ portfolio/src/components/Navbar/Navbar.js
@@ -5,7 +5,6 @@ import Burger from './Burger';
 const Nav = styled.nav`
   width: 97%;
   height: 55px;
-  padding: 0 20px;
   display: flex;
   justify-content: space-between;
